+++ src/fabric/_types.ts
@@ -60,6 +60,8 @@ export type FabricApiItemType =
 	| "ItemConnection" 						//	An Item connection.
 	| "ItemJobInstances"					//	Folder for item job instances.
 	| "ItemJobInstance"						//	An Item job instance.
+	| "ItemJobSchedules"					//	Folder for item job schedules.
+	| "ItemJobSchedule"						//	An Item job schedule.
 	| "ItemDataAccessRoles"					//	Folder for item data access roles.	
 	| "ItemDataAccessRole"					//	An Item data access role.
 	| "WorkspaceRoleAssignments"			//	Folder for workspace role assignments.
@@ -289,25 +291,27 @@ export enum iFabricApiConnectionRoleAssignmentRole {
 	"UserWithReshare" // Enables user with resharing access for the connection.
 }
 
+export interface iFabricApiGenericPrincipal {
+	displayName: string; // The principal's display name.
+	id: string; // The principal's ID.
+	type: "User" | "Group" | "ServicePrincipal" | "ServicePrincipalProfile"; // The type of the principal. Additional principal types may be added over time.
+	servicePrincipalDetails: {
+		aadAppId: string; // The service principal ID of the principal.
+	}; // Service principal specific details. Applicable when the principal type is ServicePrincipal.
+	servicePrincipalProfileDetails: {
+		parentPrincipal: object; // The service principal profile ID of the principal.
+	}; // Service principal profile details. Applicable when the principal type is ServicePrincipalProfile.
+	userDetails: {
+		userPrincipalName: string; // The user principal name of the principal.
+	}; // The user details of the principal.
+	groupDetails: {
+		groupType: string
+	};
+}
+
 export interface iFabricApiGenericRoleAssignment {
 	id: string; // The unique id for the role assignment.
-	principal: {
-		displayName: string; // The principal's display name.
-		id: string; // The principal's ID.
-		type: "User" | "Group" | "ServicePrincipal" | "ServicePrincipalProfile"; // The type of the principal. Additional principal types may be added over time.
-		servicePrincipalDetails: {
-			aadAppId: string; // The service principal ID of the principal.
-		}; // Service principal specific details. Applicable when the principal type is ServicePrincipal.
-		servicePrincipalProfileDetails: {
-			parentPrincipal: object; // The service principal profile ID of the principal.
-		}; // Service principal profile details. Applicable when the principal type is ServicePrincipalProfile.
-		userDetails: {
-			userPrincipalName: string; // The user principal name of the principal.
-		}; // The user details of the principal.
-		groupDetails: {
-			groupType: string
-		};
-	}; // The principal object which contains the principal details.
+	principal: iFabricApiGenericPrincipal; // The principal object which contains the principal details.
 	role: any; // The role of the role assignment.
 }
 
@@ -336,12 +340,20 @@ export interface iFabricApiItemJobInstance {
 	status: string; // The item job status. Additional statuses may be added over time.
 }
 
+export interface iFabricApiItemJobSchedule {
+	configuration: any; // The actual data contains the time/weekdays of this schedule.
+	createdDateTime: string; // The created time stamp of this schedule in Utc.
+	enabled: boolean; // Whether this schedule is enabled. True - Enabled, False - Disabled.
+	id: string; // The schedule ID.
+	owner: iFabricApiGenericPrincipal; // The user identity that created this schedule or last modified.
+}
+
 export interface iFabricApiGateway {
 	displayName: string; // The display name of the on-premises gateway.
 	id: string; // The object ID of the gateway.
 	numberOfMemberGateways: number; // The number of gateway members in the on-premises gateway.
 	type: "OnPremises" | "OnPremisesPersonal" | "VirtualNetwork"; // The type of the gateway.
-	
+
 	// on-premises + personal gateway only
 	publicKey: any; // The public key of the primary gateway member. Used to encrypt the credentials for creating and updating connections.
 	version: string; // The version of the installed primary gateway member.
@@ -350,7 +362,7 @@ export interface iFabricApiGateway {
 	allowCloudConnectionRefresh: boolean; // Whether to allow cloud connections to refresh through this on-premises gateway. True - Allow, False - Do not allow.
 	allowCustomConnectors: boolean; // Whether to allow custom connectors to be used with this on-premises gateway. True - Allow, False - Do not allow.
 	loadBalancingSetting: any; // The load balancing setting of the on-premises gateway.
-	
+
 
 	// VNet Gateway only
 	capacityId: string; // The object ID of the Fabric license capacity.